+++ src/named.rs
@@ -224,11 +224,7 @@ pub enum NamedChain {
 
     Shimmer = 148,
 
-
-    #[cfg_attr(
-        feature = "serde",
-        serde(alias = "ink_sepolia_testnet", alias = "ink-sepolia")
-    )]
+    #[cfg_attr(feature = "serde", serde(alias = "ink_sepolia_testnet", alias = "ink-sepolia"))]
     InkSepolia = 763373,
 
     #[strum(to_string = "fraxtal")]
@@ -381,6 +377,13 @@ pub enum NamedChain {
     #[strum(to_string = "unichain-sepolia")]
     #[cfg_attr(feature = "serde", serde(alias = "unichain-sepolia"))]
     UnichainSepolia = 1301,
+
+    #[strum(to_string = "apechain")]
+    #[cfg_attr(feature = "serde", serde(alias = "apechain"))]
+    ApeChain = 33139,
+    #[strum(to_string = "curtis", serialize = "apechain-testnet")]
+    #[cfg_attr(feature = "serde", serde(alias = "apechain-testnet", alias = "curtis"))]
+    Curtis = 33111,
 }
 
 // This must be implemented manually so we avoid a conflict with `TryFromPrimitive` where it treats
@@ -575,11 +578,13 @@ impl NamedChain {
             | XaiSepolia
             | Syndr
             | SyndrSepolia
-            | ArbitrumNova => 260,
+            | ArbitrumNova
+            | ApeChain
+            | Curtis => 260,
 
-            Optimism | OptimismGoerli | OptimismSepolia | Base | BaseGoerli | BaseSepolia | InkSepolia
-            | Blast | BlastSepolia | Fraxtal | FraxtalTestnet | Zora | ZoraGoerli | ZoraSepolia
-            | Mantle | MantleSepolia | Mode | ModeSepolia | Pgn | PgnSepolia
+            Optimism | OptimismGoerli | OptimismSepolia | Base | BaseGoerli | BaseSepolia
+            | InkSepolia | Blast | BlastSepolia | Fraxtal | FraxtalTestnet | Zora | ZoraGoerli
+            | ZoraSepolia | Mantle | MantleSepolia | Mode | ModeSepolia | Pgn | PgnSepolia
             | SoneiumMinatoTestnet => 2_000,
 
             Odyssey => 1_000,
@@ -776,7 +781,9 @@ impl NamedChain {
             | SoneiumMinatoTestnet
             | World
             | WorldSepolia
-            | UnichainSepolia => false,
+            | UnichainSepolia
+            | ApeChain
+            | Curtis => false,
 
             // Unknown / not applicable, default to false for backwards compatibility.
             Dev | AnvilHardhat | Morden | Ropsten | Rinkeby | Cronos | CronosTestnet | Kovan
@@ -860,6 +867,8 @@ impl NamedChain {
                 | WorldSepolia
                 | Iotex
                 | UnichainSepolia
+                | ApeChain
+                | Curtis
         )
     }
 
@@ -934,7 +943,8 @@ impl NamedChain {
             | ImmutableTestnet
             | SoneiumMinatoTestnet
             | WorldSepolia
-            | UnichainSepolia => true,
+            | UnichainSepolia
+            | Curtis => true,
 
             // Dev chains.
             Dev | AnvilHardhat => true,
@@ -947,7 +957,7 @@ impl NamedChain {
             | Fraxtal | Linea | ZkSync | Mantle | GravityAlphaMainnet | Xai | Zora | Pgn | Mode
             | Viction | Elastos | Degen | OpBNBMainnet | Ronin | Taiko | Flare | Acala | Karura
             | Darwinia | Cfx | Crab | Pulsechain | Etherlink | Immutable | World | Iotex | Core
-            | Merlin | Bitlayer => false,
+            | Merlin | Bitlayer | ApeChain => false,
         }
     }
 
@@ -997,6 +1007,8 @@ impl NamedChain {
             Core => "CORE",
             Merlin => "BTC",
             Bitlayer => "BTC",
+
+            ApeChain | Curtis => "APE",
 
             _ => return None,
         })
@@ -1124,9 +1136,10 @@ impl NamedChain {
                 ("https://api-sepolia.scrollscan.com/api", "https://sepolia.scrollscan.com")
             }
 
-            InkSepolia => {
-                ("https://explorer-sepolia.inkonchain.com/api", "https://explorer-sepolia.inkonchain.com")
-            }
+            InkSepolia => (
+                "https://explorer-sepolia.inkonchain.com/api",
+                "https://explorer-sepolia.inkonchain.com",
+            ),
 
             Shimmer => {
                 ("https://explorer.evm.shimmer.network/api", "https://explorer.evm.shimmer.network")
@@ -1330,6 +1343,9 @@ impl NamedChain {
             Core => ("https://scan.coredao.org", "https://openapi.coredao.org/api"),
             Merlin => ("https://scan.merlinchain.io", "https://scan.merlinchain.io/api"),
             Bitlayer => ("https://www.btrscan.com", "https://api.btrscan.com/scan/api"),
+
+            ApeChain => ("https://api.apescan.io/api", "https://apescan.io"),
+            Curtis => ("https://curtis.explorer.caldera.xyz/api/v2", "https://curtis.apescan.io"),
         })
     }
 
@@ -1395,7 +1411,8 @@ impl NamedChain {
             | ScrollSepolia
             | Taiko
             | TaikoHekla
-            | UnichainSepolia => "ETHERSCAN_API_KEY",
+            | UnichainSepolia
+            | ApeChain => "ETHERSCAN_API_KEY",
 
             Avalanche | AvalancheFuji => "SNOWTRACE_API_KEY",
 
@@ -1411,7 +1428,7 @@ impl NamedChain {
             | EtherlinkTestnet | Flare | FlareCoston2 | KakarotSepolia | Karura | KaruraTestnet
             | Mode | ModeSepolia | Pgn | PgnSepolia | Shimmer | Zora | ZoraGoerli | ZoraSepolia
             | Darwinia | Crab | Koi | Immutable | ImmutableTestnet | SoneiumMinatoTestnet
-            | World | WorldSepolia | InkSepolia => "BLOCKSCOUT_API_KEY",
+            | World | WorldSepolia | Curtis | InkSepolia => "BLOCKSCOUT_API_KEY",
 
             Boba => "BOBASCAN_API_KEY",
 
@@ -1535,6 +1552,7 @@ impl NamedChain {
             Core => address!("40375C92d9FAf44d2f9db9Bd9ba41a3317a2404f"),
             Merlin => address!("F6D226f9Dc15d9bB51182815b320D3fBE324e1bA"),
             Bitlayer => address!("ff204e2681a6fa0e2c3fade68a1b28fb90e4fc5f"),
+            ApeChain => address!("48b62137EdfA95a428D35C09E44256a739F6B557"),
             _ => return None,
         };
 
@@ -1624,6 +1642,8 @@ mod tests {
             (Immutable, &["immutable"]),
             (ImmutableTestnet, &["immutable-testnet"]),
             (SoneiumMinatoTestnet, &["soneium-minato-testnet"]),
+            (ApeChain, &["apechain"]),
+            (Curtis, &["apechain-testnet", "curtis"]),
         ];
 
         for &(chain, aliases) in ALIASES {