+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.43](https://github.com/alloy-rs/chains/releases/tag/v0.1.43) - 2024-10-29
+
+### Other
+
+- Add ApeChain and Curtis Testnet ([#106](https://github.com/alloy-rs/chains/issues/106))
+
 ## [0.1.42](https://github.com/alloy-rs/chains/releases/tag/v0.1.42) - 2024-10-25
+
+### Miscellaneous Tasks
+
+- Release 0.1.42
 
 ### Other